[Oztechan/CCC#2613] Move SelectCurrencyView to SelectCurrencyRootView
